+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for **response_format** parameter in completions
+- Parameter `$role` for `CompletionInterface::getLastMessage()` method, to filter last message for given role
 
 ## [1.1.0] - 2025-06-17